"optimize" rectangle pocket  cuts
since we’re clearing away a rectangle, we don’t have to move to the
safe area before movement; we can directly move in a single linear move
to a point slightly above the new cutting point, the plunge. We go
slightly above to not to leave marks while traveling.