Don't round corners for obtuse angles
For point paths which specify a lot of very close points, the small
arcs can cause a grbl board to stutter. In this situation, the angles
are all obtuse and there is extremely little benefit for using arcs.

For background: acute angles require arc cuts because cutting to the
intersection point would cut far beyond the material boundaries.