Fix #61 (sort of)
With this change, an extremely bad input (like doubling back) will result in a more reasonable response. The output will still be a bit incorrect in the sense that an inner-profile-cut probably should not even attempt to cut along the line. With this change, a warning will be emitted, and the outside of the line will be cut. I believe this is at least a reasonable response to an unreasonable input.

NOTE: There may be similar cases with very tight angles. I didn't test those.